services/langgraph-python/requirements.txt
langgraph>=0.0.38
langchain>=0.1.0
langchain-anthropic>=0.1.0
anthropic>=0.18.0
supabase>=2.3.0
reportlab>=4.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
qrcode>=7.4.2
pillow>=10.2.0
lxml>=5.1.0
cryptography>=42.0.0
